fix database migrations

+++ src/database/migrations/2020_11_16_091608_create_blog_posts_table.php
@@ -14,10 +14,10 @@ class CreateBlogPostsTable extends Migration
     public function up()
     {
         Schema::create('blog_posts', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
-            $table->bigIncrements('category_id')->unsigned();
-            $table->bigIncrements('user_id')->unsigned();
+            $table->bigInteger('category_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
 
             $table->string('slug')->unique();
             $table->string('title');
